Add functions for getting interfaces via the ament index
This is a direct copy of the definitions found in the ros2interface package:

https://github.com/ros2/ros2cli/blob/8ece0f10feba00f596514185a02cc293266e3e20/ros2interface/ros2interface/api/__init__.py

Since these functions are specific to rosidl, this package seems like a more appropriate place for them to live.
This will also let them be easier reused by several of the ros2cli packages (and potentially elsewhere in the future).

Signed-off-by: Jacob Perron <jacob@openrobotics.org>